Cleaning up pem parsing for ed25519 and removing unused hasher for secp256k1
